+++ spring-boot-mysql-springdatajpa-hibernate/src/main/java/netgloo/controllers/RegisterController.java
@@ -4,7 +4,6 @@ import netgloo.models.User;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 /**
