+++ spring-boot-mysql-springdatajpa-hibernate/src/main/java/netgloo/controllers/ShopingCartController.java
@@ -1,11 +1,8 @@
 package netgloo.controllers;
 
-import netgloo.models.Bild;
-import netgloo.models.DisplayObjects.ShoppingCartItem;
 import netgloo.models.DisplayObjects.ShoppingChart;
 import netgloo.models.daos.BildDao;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.PathVariable;
