--- spring-boot-mysql-springdatajpa-hibernate/src/main/java/netgloo/models/Authority.java
@@ -1,16 +1,0 @@
-package netgloo.models;
-
-import org.springframework.security.core.GrantedAuthority;
-
-public class Authority implements GrantedAuthority{
-	private final String authority;
-	
-	public Authority(String authority) {
-		this.authority = authority;
-	}
-	
-	@Override
-	public String getAuthority() {
-		return authority;
-	}
-}
