+++ spring-boot-mysql-springdatajpa-hibernate/src/main/java/netgloo/models/Role.java
@@ -1,43 +1,12 @@
 package netgloo.models;
 
-import javax.persistence.*;
-import java.util.HashSet;
-import java.util.Set;
+/**
+ * Created by mazi on 19.04.17.
+ */
+public enum Role {
+    USER,ADMIN,FOTOGRAF;
 
-@Entity
-public class Role {
+    Role() {
+    }
 
-	@Id
-	@GeneratedValue
-	private int roleId;
-	private String name;
-	
-	@OneToMany(mappedBy = "role", cascade=CascadeType.ALL, fetch=FetchType.LAZY)
-	private Set<UserRole> userRoles = new HashSet<>();
-
-	public int getRoleId() {
-		return roleId;
-	}
-
-	public void setRoleId(int roleId) {
-		this.roleId = roleId;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public Set<UserRole> getUserRoles() {
-		return userRoles;
-	}
-
-	public void setUserRoles(Set<UserRole> userRoles) {
-		this.userRoles = userRoles;
-	}
-	
-	
 }
