--- spring-boot-mysql-springdatajpa-hibernate/src/main/java/netgloo/models/Roles.java
@@ -1,12 +1,0 @@
-package netgloo.models;
-
-/**
- * Created by mazi on 19.04.17.
- */
-public enum Roles {
-    USER,ADMIN,FOTOGRAF;
-
-    Roles() {
-    }
-
-}
