--- spring-boot-mysql-springdatajpa-hibernate/src/main/java/netgloo/models/daos/RoleDao.java
@@ -1,11 +1,0 @@
-package netgloo.models.daos;
-
-import netgloo.models.Role;
-import org.springframework.data.repository.CrudRepository;
-
-/**
- * Created by mazi on 31.01.17.
- */
-public interface RoleDao extends CrudRepository <Role,Long> {
-
-}
