--- spring-boot-mysql-springdatajpa-hibernate/src/main/java/netgloo/models/daos/UserRoleDao.java
@@ -1,11 +1,0 @@
-package netgloo.models.daos;
-
-import netgloo.models.UserRole;
-import org.springframework.data.repository.CrudRepository;
-
-/**
- * Created by mazi on 31.01.17.
- */
-public interface UserRoleDao extends CrudRepository<UserRole,Long> {
-
-}
