+++ spring-boot-mysql-springdatajpa-hibernate/src/main/java/netgloo/models/security/UserDetailsServiceImpl.java
@@ -1,18 +1,13 @@
 package netgloo.models.security;
 
-import netgloo.models.Authority;
-import netgloo.models.Role;
 import netgloo.models.User;
 import netgloo.models.daos.UserDao;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Collection;
 
 
 /**
@@ -32,9 +27,7 @@ public class UserDetailsServiceImpl implements UserDetailsService {
         User user = userDao.findByUsername(username);
 
 
-        Collection<? extends GrantedAuthority> authorities = null;
-        authorities = user.getAuthorities();
-        return new org.springframework.security.core.userdetails.User(user.getUsername(), user.getPassword(), user.getAuthorities());
+        return null;
     }
