--- spring-boot-mysql-springdatajpa-hibernate/src/main/java/netgloo/old/UserDetailController.java
@@ -1,12 +1,0 @@
-package netgloo.old;
-
-import org.springframework.stereotype.Controller;
-
-/**
- * Created by mazi on 13.01.17.
- */
-
-public class UserDetailController {
-
-
-}
