--- spring-boot-mysql-springdatajpa-hibernate/src/main/resources/templates/old/Registrierung.html
@@ -1,99 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org" xmlns:sec="http://www.thymeleaf.org/extras/spring-security">
-
-<head>
-
-    <meta charset="utf-8"/>
-    <meta http-equiv="X-UA-Compatible" content="IE=edge"/>
-    <meta name="viewport" content="width=device-width, initial-scale=1"/>
-    <meta name="description" content=""/>
-    <meta name="author" content=""/>
-
-    <title>New Age - Start Bootstrap Theme</title>
-
-    <!-- Bootstrap Core CSS -->
-    <link href="vendor/bootstrap/css/bootstrap.min.css" rel="stylesheet"
-          th:href="@{/css/vendor/bootstrap/css/bootstrap.min.css}"/>
-
-    <!-- Custom Fonts -->
-    <link href="https://fonts.googleapis.com/css?family=Lato" rel="stylesheet"/>
-    <link href="https://fonts.googleapis.com/css?family=Catamaran:100,200,300,400,500,600,700,800,900"
-          rel="stylesheet"/>
-    <link href="https://fonts.googleapis.com/css?family=Muli" rel="stylesheet"/>
-
-    <!-- Plugin CSS -->
-    <link rel="stylesheet" th:href="@{/css/vendor/font-awesome/css/font-awesome.min.css}"/>
-    <link rel="stylesheet" th:href="@{/css/vendor/simple-line-icons/css/simple-line-icons.css}"/>
-    <link rel="stylesheet" th:href="@{/css/vendor/device-mockups/device-mockups.min.css}"/>
-
-    <!-- Theme CSS -->
-    <link href="css/new-age.min.css" rel="stylesheet" th:href="@{/css/new-age.min.css}"/>
-
-    <!-- HTML5 Shim and Respond.js IE8 support of HTML5 elements and media queries -->
-    <!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
-    <!--[if lt IE 9]>
-    <script src="https://oss.maxcdn.com/libs/html5shiv/3.7.0/html5shiv.js"></script>
-    <script src="https://oss.maxcdn.com/libs/respond.js/1.4.2/respond.min.js"></script>
-    <![endif]-->
-
-    <link rel="stylesheet" th:href="@{/css/custom.css}"/>
-</head>
-
-<body id="page-top">
-
-<div class="container" th:replace="fragments/navbar :: navbar">
-    <p>navbar</p>
-</div>
-
-<div id="register" class="animate form">
-    <form class="form-horizontal" th:object="${user,adresse}" th:action="@{/register}" method="post" autocomplete="on">
-        <h1> Login </h1>
-        <p>
-            <label for="usernamesignup" class="uname" > Benutzername</label>
-            <input id="usernamesignup" name="usernamesignup" required="required" th:field="*{user.username}" type="text" placeholder="username" />
-        </p>
-        <p>
-            <label for="emailsignup" class="youmail"  >Email</label>
-            <input id="emailsignup" name="emailsignup" required="required" th:field="*{user.email}" type="email" placeholder="mysupermail@mail.com"/>
-        </p>
-        <p>
-            <label for="passwordsignup" class="youpasswd" >Passwort</label>
-            <input id="passwordsignup" name="passwordsignup" required="required" th:field="*{user.password}" type="password" placeholder="*********"/>
-        </p>
-        <p>
-            <label for="passwordsignup_confirm" class="youpasswd" data-icon="p">Passwort wiederholen</label>
-            <input id="passwordsignup_confirm" name="passwordsignup_confirm" required="required" type="password" placeholder="**********"/>
-        </p>
-        <p class="signin button">
-            <input type="submit" value="Sign up"/>
-        </p>
-        <p class="change_link">
-            Already a member ?
-            <a href="/login" class="to_register"> Go and log in </a>
-        </p>
-    </form>
-</div>
-
-<footer>
-    <div class="container" th:replace="fragments/footer :: footer">
-        <p>footer</p>
-    </div>
-</footer>
-
-<!-- jQuery -->
-<script th:src="@{/js/jquery/jquery.min.js}"></script>
-
-<!-- Bootstrap Core JavaScript -->
-<script th:src="@{/js/bootstrap/js/bootstrap.min.js}"></script>
-
-<!-- Plugin JavaScript -->
-<script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-easing/1.3/jquery.easing.min.js"></script>
-
-<!-- Theme JavaScript -->
-<script th:src="@{/js/new-age.min.js}"></script>
-
-</body>
-
-</html>
-
-
