--- spring-boot-mysql-springdatajpa-hibernate/src/main/resources/templates/old/checkoutcomplete.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8"/>
-
-    <link href="../static/css/bootstrap-3.3.4-dist/css/bootstrap.min.css" rel="stylesheet" media="screen"
-          th:href="@{/css/bootstrap-3.3.4-dist/css/bootstrap.min.css}"/>
-
-    <script src="../../static/js/jquery-2.1.4.min.js"
-            th:src="@{/js/jquery-2.1.4.min.js}"></script>
-
-    <link th:href="@{/css/custom.css}" rel="stylesheet" media="screen"/>
-    <title>Checkout Complete</title>
-</head>
-<body>
-<div class="container">
-    <div class="row col-md-12">
-        <h2>Thank You for Your Order!!</h2>
-    </div>
-</div>
-
-</body>
-</html>
