--- spring-boot-mysql-springdatajpa-hibernate/src/main/resources/templates/old/checkoutform.html
@@ -1,122 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8"/>
-
-    <link href="../static/css/bootstrap-3.3.4-dist/css/bootstrap.min.css" rel="stylesheet" media="screen"
-          th:href="@{/css/bootstrap-3.3.4-dist/css/bootstrap.min.css}"/>
-
-    <script src="../../static/js/jquery-2.1.4.min.js"
-            th:src="@{/js/jquery-2.1.4.min.js}"></script>
-
-    <link th:href="@{/css/custom.css}" rel="stylesheet" media="screen"/>
-    <title>Checkout</title>
-</head>
-<body>
-<div class="container">
-    <h2>Checkout Form</h2>
-
-    <form class="form-horizontal" th:object="${checkoutCommand}" th:action="@{/docheckout}" method="post">
-
-        <div th:if="${#fields.hasErrors('*')}" class="alert alert-danger">
-            <p th:text="#{checkoutCommand.hasErrors}">Error Message</p>
-        </div>
-
-        <div class="form-group" th:class="${#fields.hasErrors('firstName')} ? 'form-group has-error' : 'form-group'">
-            <label class="col-sm-2 control-label">First Name:</label>
-            <div class="col-sm-10">
-                <input type="text" class="form-control" th:field="*{firstName}" th:errorclass="has-error"/>
-
-                <span class="help-block" th:if="${#fields.hasErrors('firstName')}">
-                       <ul>
-                           <li th:each="err : ${#fields.errors('firstName')}" th:text="${err}" />
-                       </ul>
-                </span>
-            </div>
-        </div>
-
-        <div class="form-group" th:class="${#fields.hasErrors('lastName')} ? 'form-group has-error' : 'form-group'">
-            <label class="col-sm-2 control-label">Last Name:</label>
-            <div class="col-sm-10">
-                <input type="text" class="form-control" th:field="*{lastName}" th:errorclass="has-error"/>
-
-                <span class="help-block" th:if="${#fields.hasErrors('lastName')}">
-                       <ul>
-                           <li th:each="err : ${#fields.errors('lastName')}" th:text="${err}" />
-                       </ul>
-                </span>
-            </div>
-        </div>
-
-        <div class="form-group" th:class="${#fields.hasErrors('addressLine1')} ? 'form-group has-error' : 'form-group'">
-            <label class="col-sm-2 control-label">Address Line 1:</label>
-            <div class="col-sm-10">
-                <input type="text" class="form-control" th:field="*{addressLine1}" th:errorclass="has-error"/>
-
-                <span class="help-block" th:if="${#fields.hasErrors('addressLine1')}">
-                       <ul>
-                           <li th:each="err : ${#fields.errors('addressLine1')}" th:text="${err}" />
-                       </ul>
-                </span>
-            </div>
-        </div>
-
-        <div class="form-group" th:class="${#fields.hasErrors('addressLine2')} ? 'form-group has-error' : 'form-group'">
-            <label class="col-sm-2 control-label">Address Line 2:</label>
-            <div class="col-sm-10">
-                <input type="text" class="form-control" th:field="*{addressLine2}" th:errorclass="has-error"/>
-
-                <span class="help-block" th:if="${#fields.hasErrors('addressLine2')}">
-                       <ul>
-                           <li th:each="err : ${#fields.errors('addressLine2')}" th:text="${err}" />
-                       </ul>
-                </span>
-            </div>
-        </div>
-
-        <div class="form-group" th:class="${#fields.hasErrors('city')} ? 'form-group has-error' : 'form-group'">
-            <label class="col-sm-2 control-label">City:</label>
-            <div class="col-sm-10">
-                <input type="text" class="form-control" th:field="*{city}" th:errorclass="has-error"/>
-
-                <span class="help-block" th:if="${#fields.hasErrors('city')}">
-                       <ul>
-                           <li th:each="err : ${#fields.errors('city')}" th:text="${err}" />
-                       </ul>
-                </span>
-            </div>
-        </div>
-
-        <div class="form-group" th:class="${#fields.hasErrors('stateCode')} ? 'form-group has-error' : 'form-group'">
-            <label class="col-sm-2 control-label">State Code:</label>
-            <div class="col-sm-10">
-                <input type="text" class="form-control" th:field="*{stateCode}" th:errorclass="has-error"/>
-
-                <span class="help-block" th:if="${#fields.hasErrors('stateCode')}">
-                       <ul>
-                           <li th:each="err : ${#fields.errors('stateCode')}" th:text="${err}" />
-                       </ul>
-                </span>
-            </div>
-        </div>
-
-        <div class="form-group" th:class="${#fields.hasErrors('zipCode')} ? 'form-group has-error' : 'form-group'">
-            <label class="col-sm-2 control-label">Zip Code:</label>
-            <div class="col-sm-10">
-                <input type="text" class="form-control" th:field="*{zipCode}" th:errorclass="has-error"/>
-
-                <span class="help-block" th:if="${#fields.hasErrors('zipCode')}">
-                       <ul>
-                           <li th:each="err : ${#fields.errors('zipCode')}" th:text="${err}" />
-                       </ul>
-                </span>
-            </div>
-        </div>
-
-        <div class="row">
-            <button type="submit" class="btn btn-default">Submit</button>
-        </div>
-    </form>
-</div>
-</body>
-</html>
