--- spring-boot-mysql-springdatajpa-hibernate/src/main/resources/templates/old/fragments/loginForm.html
@@ -1,78 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org" xmlns:sec="http://www.thymeleaf.org/extras/spring-security">
-<head>
-    <meta charset="UTF-8"/>
-    <title th:text="#{index.page.title}">Page Title</title>
-
-    <link href="vendor/bootstrap/css/bootstrap.min.css" rel="stylesheet"
-          th:href="@{/css/vendor/bootstrap/css/bootstrap.min.css}"/>
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"/>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
-    <link rel="stylesheet" th:href="@{/css/custom.css}"/>
-</head>
-<body>
-<div class="container" th:replace="fragments/navbar :: navbar">
-    <p>navbar</p>
-</div>
-<!-- Modal -->
-<div th:fragment="loginForm">
-    <div class="modal-dialog center_div" >
-
-        <!-- Modal content-->
-        <div class="modal-content ">
-            <div class="modal-header">
-                <h4 class="modal-title black">Login</h4>
-            </div>
-            <div class="modal-body">
-                <form class="form-horizontal" th:object="${loginCommand}" th:action="@{/login}" method="post">
-
-                    <div th:if="${#fields.hasErrors('*')}" class="alert alert-danger">
-                        <p th:text="#{loginCommand.hasErrors}">Error Message</p>
-                    </div>
-
-                    <div th:if="${param.error}" class="alert alert-danger">
-                        <p>Ungültiges Passwort.</p>
-                    </div>
-
-                    <div th:if="${param.logout}" class="alert alert-success">
-                        <p>Ausgelogged</p>
-                    </div>
-
-                    <div class="form-group"
-                         th:class="${#fields.hasErrors('username')} ? 'form-group has-error' : 'form-group'">
-                        <label class="col-sm-2 control-label black">Username:</label>
-                        <div class="col-sm-10">
-                            <input type="text" class="form-control" name="username" id="username"/>
-                            <span class="help-block" th:if="${#fields.hasErrors('username')}">
-                                <ul>
-                                    <li th:each="err : ${#fields.errors('username')}" th:text="${err}"/>
-                                </ul>
-                            </span>
-                        </div>
-                    </div>
-
-                    <div class="form-group"
-                         th:class="${#fields.hasErrors('password')} ? 'form-group has-error' : 'form-group'">
-                        <label class="col-sm-2 control-label black">Passwort:</label>
-                        <div class="col-sm-10">
-                            <input type="password" class="form-control" name="password" id="password"/>
-                            <span class="help-block" th:if="${#fields.hasErrors('password')}">
-                                <ul>
-                                    <li th:each="err : ${#fields.errors('password')}" th:text="${err}"/>
-                                </ul>
-                            </span>
-                        </div>
-                    </div>
-
-                    <button type="submit" class="btn btn-default">Submit</button>
-                </form>
-
-            </div>
-        </div>
-
-    </div>
-</div>
-
-</body>
-</html>
