--- spring-boot-mysql-springdatajpa-hibernate/src/main/resources/templates/old/fragments/navbar.html
@@ -1,55 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org" xmlns:sec="http://www.thymeleaf.org/extras/spring-security">
-<head>
-    <meta charset="UTF-8"/>
-    <title th:text="#{index.page.title}">Page Title</title>
-
-    <link href="vendor/bootstrap/css/bootstrap.min.css" rel="stylesheet" th:href="@{/css/vendor/bootstrap/css/bootstrap.min.css}"/>
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"/>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
-    <link rel="stylesheet" th:href="@{/css/custom.css}"/>
-</head>
-<body>
-<div th:fragment="navbar">
-<nav id="mainNav" class="navbar navbar-default navbar-fixed-top" >
-    <div class="container">
-        <!-- Brand and toggle get grouped for better mobile display -->
-        <div class="navbar-header">
-            <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#bs-example-navbar-collapse-1">
-                <span class="sr-only">Toggle navigation</span> Menu <i class="fa fa-bars"></i>
-            </button>
-            <a class="navbar-brand page-scroll" href="/">SkyPic</a>
-        </div>
-
-        <!-- Collect the nav links, forms, and other content for toggling -->
-        <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
-            <ul class="nav navbar-nav navbar-right">
-                <li>
-                    <a class="page-scroll" href="#download">Shop</a>
-                </li>
-                <li>
-                    <a class="page-scroll" href="#features">Service</a>
-                </li>
-                <li>
-                    <a class="page-scroll" href="#contact">Kontakt</a>
-                </li>
-                <li>
-                    <a href="/loginform"  sec:authorize="!isAuthenticated()">Login</a>
-                    <!--<p sec:authentication="principal.username"></p>-->
-                </li>
-                <li>
-                    <a href="/logout" sec:authorize="isAuthenticated()">Logout</a>
-                </li>
-            </ul>
-        </div>
-        <!-- /.navbar-collapse -->
-    </div>
-    <!-- /.container-fluid -->
-</nav>
-
-    <!-- Modal -->
-
-</div>
-</body>
-</html>
