--- spring-boot-mysql-springdatajpa-hibernate/src/main/resources/templates/old/userDetail.html
@@ -1,79 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org" xmlns:sec="http://www.thymeleaf.org/extras/spring-security">
-
-<head>
-
-    <meta charset="utf-8"/>
-    <meta http-equiv="X-UA-Compatible" content="IE=edge"/>
-    <meta name="viewport" content="width=device-width, initial-scale=1"/>
-    <meta name="description" content=""/>
-    <meta name="author" content=""/>
-
-    <title>New Age - Start Bootstrap Theme</title>
-
-    <!-- Bootstrap Core CSS -->
-    <link href="vendor/bootstrap/css/bootstrap.min.css" rel="stylesheet"
-          th:href="@{/css/vendor/bootstrap/css/bootstrap.min.css}"/>
-
-    <!-- Custom Fonts -->
-    <link href="https://fonts.googleapis.com/css?family=Lato" rel="stylesheet"/>
-    <link href="https://fonts.googleapis.com/css?family=Catamaran:100,200,300,400,500,600,700,800,900"
-          rel="stylesheet"/>
-    <link href="https://fonts.googleapis.com/css?family=Muli" rel="stylesheet"/>
-
-    <!-- Plugin CSS -->
-    <link rel="stylesheet" th:href="@{/css/vendor/font-awesome/css/font-awesome.min.css}"/>
-    <link rel="stylesheet" th:href="@{/css/vendor/simple-line-icons/css/simple-line-icons.css}"/>
-    <link rel="stylesheet" th:href="@{/css/vendor/device-mockups/device-mockups.min.css}"/>
-
-    <!-- Theme CSS -->
-    <link href="css/new-age.min.css" rel="stylesheet" th:href="@{/css/new-age.min.css}"/>
-
-    <!-- HTML5 Shim and Respond.js IE8 support of HTML5 elements and media queries -->
-    <!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
-    <!--[if lt IE 9]>
-    <script src="https://oss.maxcdn.com/libs/html5shiv/3.7.0/html5shiv.js"></script>
-    <script src="https://oss.maxcdn.com/libs/respond.js/1.4.2/respond.min.js"></script>
-    <![endif]-->
-
-    <link rel="stylesheet" th:href="@{/css/custom.css}"/>
-</head>
-
-<body id="page-top">
-
-<div class="container" th:replace="fragments/navbar :: navbar">
-    <p>navbar</p>
-</div>
-<header>
-    <div class="container">
-        <div class="row center">
-            <div class="container" th:replace="fragments/loginForm :: loginForm">
-                <p>loginForm</p>
-            </div>
-        </div>
-    </div>
-
-</header>
-
-
-<footer>
-    <div class="container" th:replace="fragments/footer :: footer">
-        <p>footer</p>
-    </div>
-</footer>
-
-<!-- jQuery -->
-<script th:src="@{/js/jquery/jquery.min.js}"></script>
-
-<!-- Bootstrap Core JavaScript -->
-<script th:src="@{/js/bootstrap/js/bootstrap.min.js}"></script>
-
-<!-- Plugin JavaScript -->
-<script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-easing/1.3/jquery.easing.min.js"></script>
-
-<!-- Theme JavaScript -->
-<script th:src="@{/js/new-age.min.js}"></script>
-
-</body>
-
-</html>
